+++ client/src/app/stylesheets/NavBar.css
@@ -1,4 +1,4 @@
 .logo {
     padding: 1rem;
-    height: 60px;
+    height: 40px;
 }
